% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_nse_df}
\alias{as_nse_df}
\title{Convert \code{data.frame} to \code{nse_df}}
\usage{
as_nse_df(x)
}
\value{
a \code{data.frame} object with the class \code{nse_df}
}
\description{
Simple method to add \code{nse_df} class
to a \code{data.frame} object. The method only modifies the
class field of the object, and will fail by default.
}
