% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse_tibble.R
\name{nse-data.frame}
\alias{nse-data.frame}
\alias{nse_tbl}
\alias{nse_df}
\title{Non-Standard Evaulation Data Frame}
\usage{
nse_tbl(
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

nse_df(
  ...,
  row.names = NULL,
  check.rows = FALSE,
  check.names = TRUE,
  fix.empty.names = TRUE,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{...}{these arguments are of either the form \code{value} or \code{tag = value}.
Component names are created based on the tag (if present) or the deparsed
argument itself.}

\item{.rows}{The number of rows, useful to create a 0-column tibble or
just as an additional check.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{row.names}{NULL or a single integer or character
string specifying a column to be used as row names,
or a character or integer vector giving the row names
for the data frame.}

\item{check.rows}{if TRUE then the rows are checked for
consistency of length and names.}

\item{check.names}{logical. If TRUE then the names of the
variables in the data frame are checked to ensure that
they are syntactically valid variable names and are not
duplicated. If necessary they are adjusted (by \link{make.names})
so that they are.}

\item{fix.empty.names}{logical indicating if arguments which are
"unnamed" (in the sense of not being formally called as \code{someName = arg})
get an automatically constructed name or rather name "". Needs
to be set to FALSE even when check.names is false if "" names
should be kept.}

\item{stringsAsFactors}{logical: should character vectors be
converted to factors?}
}
\value{
a \code{data.frame} object with \code{nse_df} class
}
\description{
An additional class to extend \code{data.frame} object to allow
for NSE filtering on \code{i}, and tidyselect functionality on \code{j}.
This will help condense some code statements.
}
