% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse_tibble.R
\name{nse-extract}
\alias{nse-extract}
\alias{[.nse_df}
\alias{[<-.nse_df}
\alias{[[.nse_df}
\alias{[[<-.nse_df}
\title{Extract or Replace Parts of a Data Frame}
\usage{
\method{[}{nse_df}(x, i, j, drop = FALSE)

\method{[}{nse_df}(x, i, j) <- value

\method{[[}{nse_df}(x, i)

\method{[[}{nse_df}(x, i) <- value
}
\arguments{
\item{x}{a nse data frame.}

\item{i, j, ...}{elements to extract or replace. \code{i} is typically
evaluated with \link[rlang]{eval_tidy} and \code{j} is
evaluated with \link[tidyselect]{eval_select} prior to calling
\link[base]{NextMethod}. See details for more information.}

\item{drop}{logical. If TRUE the result is coerced to the lowest
possible dimension. The default is to drop if only one column is
left, but \emph{not} to drop if only one row is left. Default is
always set to FALSE.}

\item{value}{A suitable replacement value: it will be repeated a whole number
of times if necessary and it will be coerced. if NULL, deletes the column if a
single column is selected.}
}
\description{
Extract or replace subsets of a nse data frame.
}
\details{
For \code{[}, arguments \code{i} and \code{j} will be evaluated in the context of \code{x}.
If \code{i} is provided alone, it will act as if using \code{j}, i.e. evaluated
with \link[tidyselect]{eval_select}. When \code{j} is missing, \code{i} is
evaluated with \link[rlang]{eval_tidy} with \code{x} as the mask.
the evaluations of \code{i} and \code{j} are then passed to \link[base]{NextMethod}.
For \verb{[<-}, \code{j} will first attempt to evaluate as \link[tidyselect]{eval_select}, and if
an error is encountered, will evaluate with \link[rlang]{eval_tidy} with \code{x}
as the mask.
@return a modifed \code{data.frame} object
}
